+++ components/createTab/BottomActionsBar.tsx
@@ -1,20 +1,14 @@
-import { View, StyleSheet } from "react-native";
+import { View, StyleSheet, Image } from "react-native";
 
 import Ionicons from "@expo/vector-icons/Ionicons";
 
 import IconButton from "@/components/IconButton";
-import { sliceImage } from "@/components/createTab/utils";
-
-import { resetImage } from "@/store/slices/imageSlice";
-import store from "@/store";
+import { pickImageAsync, sliceImage } from "@/components/createTab/utils";
 
 const BottomActionsBar = () => {
   const ShareIcon = <Ionicons name="share-outline" size={24} />;
-  const ResetIcon = <Ionicons name="refresh-outline" size={24} />;
-
-  const handleReset = () => {
-    store.dispatch(resetImage());
-  };
+  const AddIcon = <Ionicons name="add-outline" size={24} />;
+  const moreIcon = <Ionicons name="ellipsis-horizontal" size={24} />;
 
   const handleShare = () => {
     sliceImage();
@@ -23,7 +17,8 @@ const BottomActionsBar = () => {
   return (
     <View style={styles.actionsBar} testID="actions-bar">
       <View style={styles.iconsRow}>
-        <IconButton icon={ResetIcon} onPress={handleReset} />
+        <IconButton icon={moreIcon} onPress={() => {}} />
+        <IconButton icon={AddIcon} onPress={pickImageAsync} />
         <IconButton icon={ShareIcon} onPress={handleShare} />
       </View>
     </View>
@@ -35,7 +30,7 @@ const styles = StyleSheet.create({
     position: "absolute",
     bottom: 0,
     width: "100%",
-    padding: 12,
+    height: 72,
   },
   iconsRow: {
     flexDirection: "row",