+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated App icon
 - Internally renamed Actions Bar to Bottom Actions Bar
 - Made Top Actions Bar transparent
+- Removed tabs and use single page design instead
+- Changed buttons shown in the bottom bar
 
 ### Removed